services/advertiser-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
# PostgreSQL 연동 추가
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.49
databases==0.8.0 